+++ packages/@webex/plugin-meetings/src/metrics/config.ts
@@ -140,8 +140,9 @@ export const eventType = {
   PSTN_AUDIO_ATTEMPT_START: 'client.pstnaudio.attempt.start',
   PSTN_AUDIO_ATTEMPT_FINISH: 'client.pstnaudio.attempt.finish',
   PSTN_AUDIO_ATTEMPT_SKIP: 'client.pstnaudio.attempt.skip',
-  MOVE_TO_BREAKOUT: 'client.breakout-session.move.request',
-  JOIN_BREAKOUT_RESPONSE: 'client.breakout-session.join.response',
+  BREAKOUT_MOVE_REQUEST: 'client.breakout-session.move.request',
+  BREAKOUT_MOVE_RESPONSE: 'client.breakout-session.move.response',
+  BREAKOUT_JOIN_RESPONSE: 'client.breakout-session.join.response',
 };
 
 export const error = {